테마 객체 분리
